python_service/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
Pillow
requests
python-dotenv
scikit-learn
deepface
tensorflow
tf-keras
mtcnn
retina-face
